Updated version of SequentialSearchCV
Thanks to @nullpunktTUD for preparing the fundament for SequentialSearchCV.

Cannot initialize properties in the constructor. TODO: Can we somehow solve this?

The results of the final optimization steps are now the default attributes and accessible via properties.

TODO: Should we add a flag to just keep the results of the final optimization step? This would make the object smaller but we cannot check plausibility of previous optimization steps.